Use MinIO container from Testcontainers
