+++ eclipsestore/src/test/groovy/io/micronaut/eclipsestore/s3/S3StorageSpec.groovy
@@ -9,13 +9,12 @@ import io.micronaut.context.annotation.Replaces
 import io.micronaut.context.annotation.Requires
 import io.micronaut.core.annotation.NonNull
 import io.micronaut.core.util.StringUtils
-import io.micronaut.inject.qualifiers.Qualifiers
 import io.micronaut.eclipsestore.BaseStorageSpec
 import io.micronaut.eclipsestore.testutils.MinioLocal
-import io.micronaut.test.extensions.spock.annotation.MicronautTest
-import io.micronaut.eclipsestore.testutils.MinioUtils
 import io.micronaut.eclipsestore.testutils.S3Configuration
 import io.micronaut.eclipsestore.testutils.S3ConfigurationProperties
+import io.micronaut.inject.qualifiers.Qualifiers
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
 import io.micronaut.test.support.TestPropertyProvider
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
@@ -90,10 +89,11 @@ class S3StorageSpec extends BaseStorageSpec implements TestPropertyProvider {
     @Factory
     @Requires(property = "spec.name", value = "S3StorageSpec")
     static class MinioClient {
+
         @Replaces(S3Client)
         @Singleton
         S3Client buildClient(S3Config s3Config) {
-            MinioUtils.s3Client(s3Config)
+            MinioLocal.s3Client(s3Config)
         }
     }
