+++ gradle/libs.versions.toml
@@ -35,6 +35,7 @@ eclipsestore-persistence-binary-jdk8 = { module = 'org.eclipse.serializer:persis
 eclipsestore-persistence-binary-jdk17 = { module = 'org.eclipse.serializer:persistence-binary-jdk17', version.ref = 'managed-eclipsestore' }
 
 testcontainers = { module = "org.testcontainers:testcontainers", version.ref = "testcontainers" }
+testcontainers-minio = { module = "org.testcontainers:minio", version.ref = "testcontainers" }
 testcontainers-junit-jupiter = { module = "org.testcontainers:junit-jupiter", version.ref = "testcontainers" }
 jupiter-jupiter-params = { module = 'org.junit.jupiter:junit-jupiter-params' }
 jupiter-api = { module = 'org.junit.jupiter:junit-jupiter-api' }
