+++ test-suite-utils/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 }
 
 dependencies {
+    implementation(libs.testcontainers.minio)
     implementation(libs.managed.eclipsestore.aws.s3)
     implementation(platform(mnAws.micronaut.aws.bom))
     implementation(libs.awssdk.s3)
