--- test-suite-utils/src/main/java/io/micronaut/eclipsestore/testutils/MinioUtils.java
@@ -1,32 +1,0 @@
-package io.micronaut.eclipsestore.testutils;
-
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.Bucket;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-
-public final class MinioUtils {
-
-    private MinioUtils() {
-    }
-
-    public static S3Client s3Client(S3ConfigurationProperties s3Config) throws URISyntaxException {
-        S3Client client = S3Client.builder()
-            .endpointOverride(new URI(s3Config.getS3Configuration().getEndpointOverride()))
-            .credentialsProvider(StaticCredentialsProvider.create(AwsBasicCredentials.create(s3Config.getAccessKeyId(), s3Config.getSecretKey())))
-            .region(Region.of(s3Config.getRegion()))
-            .serviceConfiguration(b -> b  // Required for minio
-                .checksumValidationEnabled(false)
-                .pathStyleAccessEnabled(true)
-            )
-            .build();
-        if (client.listBuckets().buckets().stream().map(Bucket::name).noneMatch(s3Config.getBucketName()::equals)) {
-            client.createBucket(b -> b.bucket(s3Config.getBucketName()));
-        }
-        return client;
-    }
-}
