+++ test-suite/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     testImplementation(libs.awssdk.dynamodb)
 
     testImplementation(libs.testcontainers)
+    testImplementation(libs.testcontainers.minio)
     testImplementation(libs.testcontainers.junit.jupiter)
     testImplementation(projects.testSuiteUtils)
 }
