+++ test-suite/src/test/java/io/micronaut/eclipsestore/docs/S3MinioClient.java
@@ -4,7 +4,7 @@ import io.micronaut.context.annotation.Factory;
 import io.micronaut.context.annotation.Replaces;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.util.StringUtils;
-import io.micronaut.eclipsestore.testutils.MinioUtils;
+import io.micronaut.eclipsestore.testutils.MinioLocal;
 import jakarta.inject.Singleton;
 import software.amazon.awssdk.services.s3.S3Client;
 
@@ -17,6 +17,6 @@ public class S3MinioClient {
     @Singleton
     @Replaces(S3Client.class)
     S3Client buildClient(S3MinioConfig s3Config) throws URISyntaxException {
-        return MinioUtils.s3Client(s3Config);
+        return MinioLocal.s3Client(s3Config);
     }
 }
